+++ nexus_reader/inc/eventDataTestHelper.h
@@ -7,13 +7,13 @@ class EventDataTestHelper {
 public:
   EventDataTestHelper(const uint8_t *);
 
-  std::vector<int32_t> getDetId() { return m_detId; };
-  std::vector<float> getTof() { return m_tof; };
+  std::vector<uint32_t> getDetId() { return m_detId; };
+  std::vector<uint64_t> getTof() { return m_tof; };
   int32_t getCount() { return m_count; };
 
 private:
-  std::vector<int32_t> m_detId;
-  std::vector<float> m_tof;
+  std::vector<uint32_t> m_detId;
+  std::vector<uint64_t> m_tof;
   int32_t m_count;
 };
